First pass at moving around tensorflow code for Differentiable conformances and derivatives into Differentiable modeule. Staging for upstreaming.
