+++ stdlib/public/Differentiation/CMakeLists.txt
@@ -21,7 +21,16 @@
 # SWIFT_ENABLE_TENSORFLOW END
 
 add_swift_target_library(swift_Differentiation ${SWIFT_STDLIB_LIBRARY_BUILD_TYPES} IS_STDLIB
-  Empty.swift
+
+  GYB_SOURCES
+    tgmathDerivatives.swift.gyb
+    SIMDVectorTypesDerivatives.swift.gyb
+
+  SWIFT_MODULE_DEPENDS_LINUX Glibc
+  SWIFT_MODULE_DEPENDS_FREEBSD Glibc
+  SWIFT_MODULE_DEPENDS_CYGWIN Glibc
+  SWIFT_MODULE_DEPENDS_HAIKU Glibc
+  SWIFT_MODULE_DEPENDS_WINDOWS MSVCRT
 
   SWIFT_COMPILE_FLAGS
     ${SWIFT_STANDARD_LIBRARY_SWIFT_FLAGS}
