+++ stdlib/public/Differentiation/FloatingPointTypesDerivatives.swift.gyb
@@ -242,3 +242,28 @@ extension ${Self} : VectorProtocol {
 #endif
 % end
 % end
+
+
+extension FloatingPoint where Self : Differentiable,
+                              Self == Self.TangentVector {
+  /// The vector-Jacobian product function of `addingProduct`. Returns the
+  /// original result and pullback of `addingProduct` with respect to `self`,
+  /// `lhs` and `rhs`.
+  @inlinable
+  @derivative(of: addingProduct)
+  func _vjpAddingProduct(
+    _ lhs: Self, _ rhs: Self
+  ) -> (value: Self, pullback: (Self) -> (Self, Self, Self)) {
+    return (addingProduct(lhs, rhs), { _ in (1, rhs, lhs) })
+  }
+
+  /// The vector-Jacobian product function of `squareRoot`. Returns the original
+  /// result and pullback of `squareRoot` with respect to `self`.
+  @inlinable // FIXME(sil-serialize-all)
+  @derivative(of: squareRoot)
+  func _vjpSquareRoot() -> (value: Self, pullback: (Self) -> Self) {
+    let y = squareRoot()
+    return (y, { v in v / (2 * y) })
+  }
+}
+
