+++ stdlib/public/Differentiation/tgmathDerivatives.swift.gyb
@@ -14,6 +14,17 @@
 // This file defines derivatives for tgmath functions.
 //===----------------------------------------------------------------------===//
 
+
+#if os(macOS) || os(iOS) || os(tvOS) || os(watchOS)
+  import Darwin.C.tgmath
+#elseif os(Linux) || os(FreeBSD) || os(PS4) || os(Android) || os(Cygwin) || os(Haiku)
+  import Glibc
+#elseif os(Windows)
+  import MSVCRT
+#else
+#error("Unsupported platform")
+#endif
+
 @usableFromInline
 @derivative(of: fma)
 func _jvpFma<T: FloatingPoint & Differentiable> (
