+++ stdlib/public/core/SIMDVectorTypes.swift.gyb
@@ -45,10 +45,6 @@ public struct SIMD${n}<Scalar>: SIMD where Scalar: SIMDScalar {
   }
 
   /// Accesses the scalar at the specified position.
-  // SWIFT_ENABLE_TENSORFLOW
-  @differentiable(
-    where Scalar : EuclideanDifferentiable & BinaryFloatingPoint,
-          Scalar.TangentVector : BinaryFloatingPoint)
   public subscript(index: Int) -> Scalar {
     @_transparent get {
       _precondition(indices.contains(index))
@@ -186,36 +182,6 @@ extension SIMD${n} where Scalar: BinaryFloatingPoint {
   where Other: BinaryFloatingPoint {
     self.init()
     for i in indices { self[i] = Scalar(other[i]) }
-  }
-}
-
-// SWIFT_ENABLE_TENSORFLOW
-extension SIMD${n} : AdditiveArithmetic where Scalar : FloatingPoint {}
-
-extension SIMD${n} : Differentiable
-  where Scalar : Differentiable & BinaryFloatingPoint,
-        Scalar.TangentVector : BinaryFloatingPoint {
-  public typealias TangentVector = SIMD${n}
-}
-
-extension SIMD${n} : EuclideanDifferentiable
-  where Scalar : EuclideanDifferentiable & BinaryFloatingPoint,
-        Scalar.TangentVector : BinaryFloatingPoint {
-}
-
-extension SIMD${n}
-  where Scalar : EuclideanDifferentiable & BinaryFloatingPoint,
-        Scalar.TangentVector : BinaryFloatingPoint {
-  // NOTE(TF-1094): serialized `@derivative` for `.swiftinterface` compilation.
-  @inlinable
-  @derivative(of: subscript(_:))
-  internal func _vjpSubscript(index: Int)
-  -> (value: Scalar, pullback: (Scalar.TangentVector) -> TangentVector) {
-    return (self[index], { v in
-      var zeros = Self.zero
-      zeros[index] = Scalar(v)
-      return zeros
-    })
   }
 }
