+++ test/AutoDiff/downstream/compiler_crashers_fixed/tf1126-derivative-generic-specialization.swift
@@ -5,6 +5,8 @@
 // Related to `@differentiable` function with `partial_apply` operands,
 // to be specialized. Occurs only with `-O`.
 
+import _Differentiation
+
 struct A: Differentiable{
   var b: SIMD8<Float>
 }
