+++ test/AutoDiff/downstream/simd.swift
@@ -2,6 +2,8 @@
 // REQUIRES: executable_test
 
 import StdlibUnittest
+import _Differentiation
+
 #if os(macOS)
 import Darwin.C
 #else
